refactor: move session detail tests into session_detail/tests.rs
Relocate the ~920-line #[cfg(test)] module out of session_detail.rs
into a child module file, shrinking the main file by ~30% with no test
changes.
